Move files to relevant directory
